+++ yandex/data_source_yandex_mdb_clickhouse_cluster.go
@@ -411,6 +411,12 @@ func dataSourceYandexMDBClickHouseCluster() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+			"security_group_ids": {
+				Type:     schema.TypeSet,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Set:      schema.HashString,
+				Computed: true,
+			},
 		},
 	}
 }
@@ -526,6 +532,10 @@ func dataSourceYandexMDBClickHouseClusterRead(d *schema.ResourceData, meta inter
 	}
 
 	if err := d.Set("labels", cluster.Labels); err != nil {
+		return err
+	}
+
+	if err := d.Set("security_group_ids", cluster.SecurityGroupIds); err != nil {
 		return err
 	}