+++ yandex/data_source_yandex_mdb_mongodb_cluster.go
@@ -223,6 +223,12 @@ func dataSourceYandexMDBMongodbCluster() *schema.Resource {
 					},
 				},
 			},
+			"security_group_ids": {
+				Type:     schema.TypeSet,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Set:      schema.HashString,
+				Computed: true,
+			},
 		},
 	}
 }
@@ -340,6 +346,10 @@ func dataSourceYandexMDBMongodbClusterRead(d *schema.ResourceData, meta interfac
 	}
 
 	if err := d.Set("labels", cluster.Labels); err != nil {
+		return err
+	}
+
+	if err := d.Set("security_group_ids", cluster.SecurityGroupIds); err != nil {
 		return err
 	}