+++ yandex/data_source_yandex_mdb_mysql_cluster_test.go
@@ -182,6 +182,10 @@ func testAccDataSourceMDBMysqlClusterAttributesCheck(datasourceName string, reso
 				"host.0.zone",
 				"host.0.zone",
 			},
+			{
+				"security_group_ids.#",
+				"security_group_ids.#",
+			},
 		}
 
 		for _, attrToCheck := range instanceAttrsToTest {
@@ -222,6 +226,7 @@ func testAccDataSourceMDBMysqlClusterCheck(datasourceName string, resourceName s
 		resource.TestCheckResourceAttr(datasourceName, "host.#", "1"),
 		resource.TestCheckResourceAttrSet(datasourceName, "host.0.fqdn"),
 		testAccCheckCreatedAtAttr(datasourceName),
+		resource.TestCheckResourceAttr(datasourceName, "security_group_ids.#", "1"),
 	)
 }