+++ yandex/data_source_yandex_mdb_postgresql_cluster.go
@@ -302,6 +302,12 @@ func dataSourceYandexMDBPostgreSQLCluster() *schema.Resource {
 					},
 				},
 			},
+			"security_group_ids": {
+				Type:     schema.TypeSet,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Set:      schema.HashString,
+				Computed: true,
+			},
 		},
 	}
 }
@@ -379,6 +385,10 @@ func dataSourceYandexMDBPostgreSQLClusterRead(d *schema.ResourceData, meta inter
 	}
 
 	if err := d.Set("labels", cluster.Labels); err != nil {
+		return err
+	}
+
+	if err := d.Set("security_group_ids", cluster.SecurityGroupIds); err != nil {
 		return err
 	}