+++ yandex/resource_yandex_mdb_mysql_cluster_test.go
@@ -117,6 +117,7 @@ func TestAccMDBMySQLCluster_full(t *testing.T) {
 					testAccCheckMDBMysqlClusterContainsLabel(&cluster, "test_key", "test_value"),
 					testAccCheckCreatedAtAttr(mysqlResource),
 					testAccCheckMDBMysqlClusterHasHosts(mysqlResource, 1),
+					resource.TestCheckResourceAttr(mysqlResource, "security_group_ids.#", "1"),
 				),
 			},
 			mdbMysqlClusterImportStep(mysqlResource),
@@ -143,6 +144,7 @@ func TestAccMDBMySQLCluster_full(t *testing.T) {
 					testAccCheckMDBMysqlClusterContainsLabel(&cluster, "new_key", "new_value"),
 					testAccCheckCreatedAtAttr(mysqlResource),
 					testAccCheckMDBMysqlClusterHasHosts(mysqlResource, 1),
+					resource.TestCheckResourceAttr(mysqlResource, "security_group_ids.#", "2"),
 				),
 			},
 			mdbMysqlClusterImportStep(mysqlResource),
@@ -165,6 +167,7 @@ func TestAccMDBMySQLCluster_full(t *testing.T) {
 					testAccCheckMDBMysqlClusterContainsLabel(&cluster, "new_key", "new_value"),
 					testAccCheckCreatedAtAttr(mysqlResource),
 					testAccCheckMDBMysqlClusterHasHosts(mysqlResource, 3),
+					resource.TestCheckResourceAttr(mysqlResource, "security_group_ids.#", "1"),
 				),
 			},
 			mdbMysqlClusterImportStep(mysqlResource),
@@ -443,6 +446,43 @@ resource "yandex_vpc_subnet" "foo_a" {
   network_id     = "${yandex_vpc_network.foo.id}"
   v4_cidr_blocks = ["10.5.0.0/24"]
 }
+
+resource "yandex_vpc_security_group" "sg-x" {
+  network_id     = "${yandex_vpc_network.foo.id}"
+  ingress {
+    protocol          = "ANY"
+    description       = "Allow incoming traffic from members of the same security group"
+    from_port         = 0
+    to_port           = 65535
+    v4_cidr_blocks    = ["0.0.0.0/0"]
+  }
+  egress {
+    protocol          = "ANY"
+    description       = "Allow outgoing traffic to members of the same security group"
+    from_port         = 0
+    to_port           = 65535
+    v4_cidr_blocks    = ["0.0.0.0/0"]
+  }
+}
+
+resource "yandex_vpc_security_group" "sg-y" {
+  network_id     = "${yandex_vpc_network.foo.id}"
+  
+  ingress {
+    protocol          = "ANY"
+    description       = "Allow incoming traffic from members of the same security group"
+    from_port         = 0
+    to_port           = 65535
+    v4_cidr_blocks    = ["0.0.0.0/0"]
+  }
+  egress {
+    protocol          = "ANY"
+    description       = "Allow outgoing traffic to members of the same security group"
+    from_port         = 0
+    to_port           = 65535
+    v4_cidr_blocks    = ["0.0.0.0/0"]
+  }
+}
 `
 
 func testAccMDBMySQLClusterConfigMain(name, desc string) string {
@@ -485,6 +525,8 @@ resource "yandex_mdb_mysql_cluster" "foo" {
     zone      = "ru-central1-c"
     subnet_id = "${yandex_vpc_subnet.foo_c.id}"
   }
+
+  security_group_ids = ["${yandex_vpc_security_group.sg-x.id}"]
 }
 `, name, desc)
 }
@@ -530,6 +572,8 @@ resource "yandex_mdb_mysql_cluster" "foo" {
     subnet_id        = "${yandex_vpc_subnet.foo_c.id}"
     assign_public_ip = true
   }
+
+  security_group_ids = ["${yandex_vpc_security_group.sg-x.id}"]
 }
 `, name, desc)
 }
@@ -594,6 +638,8 @@ resource "yandex_mdb_mysql_cluster" "foo" {
     zone      = "ru-central1-c"
     subnet_id = "${yandex_vpc_subnet.foo_c.id}"
   }
+
+  security_group_ids = ["${yandex_vpc_security_group.sg-x.id}", "${yandex_vpc_security_group.sg-y.id}"]
 }
 `, name, desc)
 }
@@ -668,6 +714,8 @@ resource "yandex_mdb_mysql_cluster" "foo" {
     zone      = "ru-central1-a"
     subnet_id = "${yandex_vpc_subnet.foo_a.id}"
   }
+
+  security_group_ids = ["${yandex_vpc_security_group.sg-y.id}"]
 }
 `, name, desc)
 }